refactor: update worker background imports and improve module structure
- Changed import statements in various worker background scripts to use the new path for worker_background_worker_minify.js.
- Updated package.json to reflect the new version of @bjorn3/browser_wasi_shim.
- Enhanced the worker background worker module by adding TypeScript definitions.
- Modified the post-build script to copy the new TypeScript definition file.
- Refactored the worker background implementation to improve memory management and worker lifecycle handling.
- Removed unnecessary comments and improved code readability in worker_blob.ts.
