+++ threads/examples/package.json
@@ -3,7 +3,7 @@
     "@oligami/shared-object": "^0.1.1",
     "@xterm/xterm": "^5.5",
     "xterm-addon-fit": "^0.8.0",
-    "@bjorn3/browser_wasi_shim": "^0.3.0",
+    "@bjorn3/browser_wasi_shim": "^0.4.2",
     "@swc-node/register": "^1.11.1",
     "@oligami/browser_wasi_shim-threads": "file:../"
   },
