+++ threads/examples/wasi_multi_threads/worker_background.ts
@@ -1,5 +1,5 @@
-// @ts-ignore
-import worker_background_worker from "../node_modules/@oligami/browser_wasi_shim-threads/dist/worker_background_worker.min.js";
+// @ts-expect-error
+import worker_background_worker from "../../../threads/src/shared_array_buffer/worker_background/worker_background_worker_minify.js";
 
 import { wait_async_polyfill } from "../../src";
