+++ threads/examples/wasi_multi_threads_channel/worker_background.ts
@@ -1,8 +1,7 @@
-// @ts-ignore
-import run from "../node_modules/@oligami/browser_wasi_shim-threads/dist/worker_background_worker.min.js";
-
+// @ts-expect-error
+import worker_background_worker from "../../../threads/src/shared_array_buffer/worker_background/worker_background_worker_minify.js";
 import { wait_async_polyfill } from "../../src";
 
 wait_async_polyfill();
 
-run();
+worker_background_worker();
