+++ threads/examples/wasi_multi_threads_channel_node/worker_background_worker.ts
@@ -1,5 +1,5 @@
-// @ts-ignore
-import worker_background_worker from "../node_modules/@oligami/browser_wasi_shim-threads/dist/worker_background_worker.min.js";
+// @ts-expect-error
+import worker_background_worker from "../../../threads/src/shared_array_buffer/worker_background/worker_background_worker_minify.js";
 
 import { set_fake_worker } from "./common.ts";
