+++ threads/src/shared_array_buffer/worker_background/index.ts
@@ -9,8 +9,8 @@ import { url as gen_worker_background_worker_url } from "./worker_blob.ts";
 import type { WorkerBackgroundRefObject } from "./worker_export.ts";
 
 export {
-  WorkerBackgroundRef,
-  WorkerRef,
-  type WorkerBackgroundRefObject,
   gen_worker_background_worker_url,
+  WorkerBackgroundRef,
+  type WorkerBackgroundRefObject,
+  WorkerRef,
 };
