+++ threads/src/shared_array_buffer/worker_background/worker_blob.ts
@@ -1,14 +1,6 @@
-/**
- * Generates a Blob URL for the background worker script.
- *
- * This function creates a script from a string and returns a URL that can be
- * used to instantiate a new Worker.
- *
- * @returns A Blob URL for the worker.
- */
 export const url = () => {
   const code =
-    'let worker_background;class AllocatorUseArrayBuffer{share_arrays_memory;constructor(e=new SharedArrayBuffer(0xa00000)){this.share_arrays_memory=e;let r=new Int32Array(this.share_arrays_memory);Atomics.store(r,0,0),Atomics.store(r,1,0),Atomics.store(r,2,12)}static init_self(e){return new AllocatorUseArrayBuffer(e.share_arrays_memory)}async async_write(e,r,t){let o=new Int32Array(this.share_arrays_memory);for(;;){let{value:s}=Atomics.waitAsync(o,0,1);if("timed-out"===(s instanceof Promise?await s:s))throw Error("timed-out lock");if(0!==Atomics.compareExchange(o,0,0,1))continue;let i=this.write_inner(e,r,t);return Atomics.store(o,0,0),Atomics.notify(o,0,1),i}}block_write(e,r,t){for(;;){let o=new Int32Array(this.share_arrays_memory);if("timed-out"===Atomics.wait(o,0,1))throw Error("timed-out lock");if(0!==Atomics.compareExchange(o,0,0,1))continue;let s=this.write_inner(e,r,t);return Atomics.store(o,0,0),Atomics.notify(o,0,1),s}}write_inner(e,r,t){let o,s,i=new Int32Array(this.share_arrays_memory),a=new Uint8Array(this.share_arrays_memory);o=0===Atomics.add(i,1,1)?Atomics.store(i,2,12):Atomics.load(i,2);let n=this.share_arrays_memory.byteLength,c=e.byteLength,l=o+c;if(n<l)throw Error("size is bigger than memory. \\nTODO! fix memory limit. support big size another way.");if(e instanceof Uint8Array)s=e;else if(e instanceof Uint32Array){let r=new ArrayBuffer(e.byteLength);new Uint32Array(r).set(e),s=new Uint8Array(r)}a.set(new Uint8Array(s),o),Atomics.store(i,2,l);let w=new Int32Array(r);return Atomics.store(w,t,o),Atomics.store(w,t+1,c),[o,c]}free(e,r){Atomics.sub(new Int32Array(this.share_arrays_memory),1,1)}get_memory(e,r){let t=new ArrayBuffer(r);return new Uint8Array(t).set(new Uint8Array(this.share_arrays_memory).slice(e,e+r)),t}use_defined_memory(e,r,t){new Uint8Array(this.share_arrays_memory).set(new Uint8Array(t).slice(0,r),e)}get_object(){return{share_arrays_memory:this.share_arrays_memory}}}let serialize=e=>({message:e.message,name:e.name,stack:e.stack,cause:e.cause}),WorkerBackground=class e{override_object;allocator;lock;signature_input;workers=[void 0];start_worker;animal_workers=new Map;constructor(e,r,t,o){this.override_object=e,this.lock=r??new SharedArrayBuffer(24),this.allocator=t??new AllocatorUseArrayBuffer(new SharedArrayBuffer(10240)),this.signature_input=o??new SharedArrayBuffer(24)}static init_self(r,t){return new e(r,t.lock,AllocatorUseArrayBuffer.init_self(t.allocator),t.signature_input)}assign_worker_id(){for(let e=1;e<this.workers.length;e++)if(void 0===this.workers[e])return e;return this.workers.push(void 0),this.workers.length-1}ref(){return{allocator:this.allocator.get_object(),lock:this.lock,signature_input:this.signature_input}}destroy(){let e=0;for(let r of this.workers)void 0!==r&&(r.terminate(),console.debug("terminate worker "+e)),e++;void 0!==this.start_worker&&(this.start_worker.terminate(),console.debug("terminate start worker")),this.start_worker=void 0,this.animal_workers.clear()}async listen(){let e=new Int32Array(this.lock);Atomics.store(e,0,0),Atomics.store(e,1,1);let r=new Int32Array(this.signature_input);for(;;){let{value:t}=Atomics.waitAsync(e,1,1);if(await t==="timed-out")throw Error("timed-out");let o=()=>{let e=Atomics.load(r,1),t=Atomics.load(r,2),o=this.allocator.get_memory(e,t);return this.allocator.free(e,t),new Worker(new TextDecoder().decode(o),{type:1===Atomics.load(r,3)?"module":"classic"})},s=()=>{let e=Atomics.load(r,4),t=Atomics.load(r,5),o=this.allocator.get_memory(e,t);return this.allocator.free(e,t),JSON.parse(new TextDecoder().decode(o))};switch(Atomics.load(r,0)){case 1:{let e=o(),t=s(),i=this.assign_worker_id();console.debug("new worker "+i),this.workers[i]=e;let{promise:a,resolve:n}=Promise.withResolvers();e.onmessage=async r=>{let{msg:t,animal_id:o}=r.data;if("ready"===t&&(void 0!==o&&this.animal_workers.set(o,e),n()),"done"===t&&(this.workers[i].terminate(),this.workers[i]=void 0,void 0!==o&&this.animal_workers.delete(o),console.debug(`worker ${i} done so terminate`)),"error"===t||"exit"===t){this.workers[i]=void 0,void 0!==o&&this.animal_workers.delete(o);let e=0;for(let r of this.workers)void 0!==r&&(r.terminate(),"error"===t?console.warn("wasi throw error but child process exists, terminate "+e):console.debug("wasi exit but child process exists, terminate "+e)),e++;void 0!==this.start_worker&&(this.start_worker.terminate(),"error"===t?console.warn("wasi throw error but wasi exists, terminate wasi"):console.debug("wasi exit but wasi exists, terminate wasi")),this.workers=[void 0],this.start_worker=void 0,this.animal_workers.clear();let s=new Int32Array(this.lock,12);if("error"===t){let e=r.data.error,t=serialize(e),[o,i]=await this.allocator.async_write(new TextEncoder().encode(JSON.stringify(t)),this.lock,4);if(0!==Atomics.compareExchange(s,0,0,1)){console.error("what happened?"),this.allocator.free(o,i);return}1!==Atomics.notify(s,0)&&(console.error(e),this.allocator.free(o,i),Atomics.store(s,0,0))}else{let e=r.data.code;if(0!==Atomics.compareExchange(s,0,0,2))return void console.error("what happened?");Atomics.store(s,1,e),0===Atomics.notify(s,0)&&Atomics.store(s,0,0)}}},e.postMessage({...this.override_object,...t,worker_id:i,worker_background_ref:this.ref()}),await a,Atomics.store(r,0,i);break}case 2:{console.debug("start worker create"),this.start_worker=o();let e=s();this.start_worker.onmessage=async e=>{let{msg:r,animal_id:t}=e.data;if("ready"===r&&void 0!==t&&this.animal_workers.set(t,this.start_worker),"done"===r){let e=0;for(let r of this.workers)void 0!==r&&(r.terminate(),console.warn("wasi done but worker exists, terminate "+e)),e++;this.start_worker.terminate(),this.start_worker=void 0,void 0!==t&&this.animal_workers.delete(t);let r=new Int32Array(this.lock,12);Atomics.store(r,0,3),0===Atomics.notify(r,0)&&Atomics.store(r,0,0),console.debug("start worker done so terminate")}if("error"===r||"exit"===r){let t=0;for(let e of this.workers)void 0!==e&&(e.terminate(),"error"===r?console.warn("wasi throw error but worker exists, terminate "+t):console.debug("wasi exit but worker exists, terminate "+t)),t++;void 0!==this.start_worker&&(this.start_worker.terminate(),"error"===r?console.warn("wasi throw error but wasi exists, terminate start worker"):console.debug("wasi exit but wasi exists, terminate start worker")),this.workers=[void 0],this.start_worker=void 0,this.animal_workers.clear();let o=new Int32Array(this.lock,12);if("error"===r){let r=e.data.error,t=serialize(r),[s,i]=await this.allocator.async_write(new TextEncoder().encode(JSON.stringify(t)),this.lock,4);if(0!==Atomics.compareExchange(o,0,0,1)){console.error("what happened?"),this.allocator.free(s,i);return}if(0===Atomics.notify(o,0))throw this.allocator.free(s,i),Atomics.store(o,0,0),r}else{let r=e.data.code;if(0!==Atomics.compareExchange(o,0,0,2))return void console.error("what happened?");Atomics.store(o,1,r),0===Atomics.notify(o,0)&&Atomics.store(o,0,0)}}},this.start_worker.postMessage({...this.override_object,...e,worker_background_ref:this.ref()});break}case 3:{console.debug("terminate all workers");let e=0;for(let r of this.workers)void 0!==r&&(r.terminate(),console.debug("terminate worker "+e)),e++;void 0!==this.start_worker&&(this.start_worker.terminate(),console.debug("terminate start worker")),this.workers=[void 0],this.start_worker=void 0;break}case 4:{let e=Atomics.load(r,1);console.debug("kill animal "+e);let t=this.animal_workers.get(e);void 0!==t?(t.postMessage({msg:"kill_animal",animal_id:e}),this.animal_workers.delete(e)):console.warn(`animal ${e} not found`);break}case 5:console.debug("destroy worker background"),this.destroy()}Atomics.store(e,1,1),Atomics.store(e,2,0),0===Atomics.notify(e,2,1)&&console.warn("notify failed, waiter is late")}}};globalThis.onmessage=e=>{let{override_object:r,worker_background_ref_object:t}=e.data;worker_background=WorkerBackground.init_self(r,t),postMessage("ready")};';
+    'let worker_background;class AllocatorUseArrayBuffer{share_arrays_memory;constructor(e=new SharedArrayBuffer(0xa00000)){this.share_arrays_memory=e;const r=new Int32Array(this.share_arrays_memory);Atomics.store(r,0,0),Atomics.store(r,1,0),Atomics.store(r,2,12)}static init_self(e){return new AllocatorUseArrayBuffer(e.share_arrays_memory)}async async_write(e,r,t){let o=new Int32Array(this.share_arrays_memory);for(;;){let{value:s}=Atomics.waitAsync(o,0,1);if("timed-out"===(s instanceof Promise?await s:s))throw Error("timed-out lock");if(0!==Atomics.compareExchange(o,0,0,1))continue;let i=this.write_inner(e,r,t);return Atomics.store(o,0,0),Atomics.notify(o,0,1),i}}block_write(e,r,t){for(;;){let o=new Int32Array(this.share_arrays_memory);if("timed-out"===Atomics.wait(o,0,1))throw Error("timed-out lock");if(0!==Atomics.compareExchange(o,0,0,1))continue;let s=this.write_inner(e,r,t);return Atomics.store(o,0,0),Atomics.notify(o,0,1),s}}write_inner(e,r,t){let o,s,i=new Int32Array(this.share_arrays_memory),a=new Uint8Array(this.share_arrays_memory);o=0===Atomics.add(i,1,1)?Atomics.store(i,2,12):Atomics.load(i,2);let n=this.share_arrays_memory.byteLength,c=e.byteLength,l=o+c;if(n<l)throw Error("size is bigger than memory. \\nTODO! fix memory limit. support big size another way.");if(e instanceof Uint8Array)s=e;else if(e instanceof Uint32Array){let r=new ArrayBuffer(e.byteLength);new Uint32Array(r).set(e),s=new Uint8Array(r)}a.set(new Uint8Array(s),o),Atomics.store(i,2,l);let w=new Int32Array(r);return Atomics.store(w,t,o),Atomics.store(w,t+1,c),[o,c]}free(e,r){Atomics.sub(new Int32Array(this.share_arrays_memory),1,1)}get_memory(e,r){let t=new ArrayBuffer(r);return new Uint8Array(t).set(new Uint8Array(this.share_arrays_memory).slice(e,e+r)),t}use_defined_memory(e,r,t){new Uint8Array(this.share_arrays_memory).set(new Uint8Array(t).slice(0,r),e)}get_object(){return{share_arrays_memory:this.share_arrays_memory}}}const serialize=e=>({message:e.message,name:e.name,stack:e.stack,cause:e.cause});let WorkerBackground=class e{override_object;allocator;lock;signature_input;workers=[void 0];listen_holder;start_worker;animal_workers=new Map;constructor(e,r,t,o){this.override_object=e,this.lock=r??new SharedArrayBuffer(24),this.allocator=t??new AllocatorUseArrayBuffer(new SharedArrayBuffer(10240)),this.signature_input=o??new SharedArrayBuffer(24),this.listen_holder=this.listen()}static init_self(r,t){return new e(r,t.lock,AllocatorUseArrayBuffer.init_self(t.allocator),t.signature_input)}assign_worker_id(){for(let e=1;e<this.workers.length;e++)if(void 0===this.workers[e])return e;return this.workers.push(void 0),this.workers.length-1}ref(){return{allocator:this.allocator.get_object(),lock:this.lock,signature_input:this.signature_input}}destroy(){let e=0;for(let r of this.workers)void 0!==r&&(r.terminate(),console.debug("terminate worker "+e)),e++;void 0!==this.start_worker&&(this.start_worker.terminate(),console.debug("terminate start worker")),this.start_worker=void 0,this.animal_workers.clear()}async listen(){let e=new Int32Array(this.lock);Atomics.store(e,0,0),Atomics.store(e,1,1);let r=new Int32Array(this.signature_input);for(;;){let{value:t}=Atomics.waitAsync(e,1,1);if(await t==="timed-out")throw Error("timed-out");let o=()=>{let e=Atomics.load(r,1),t=Atomics.load(r,2),o=this.allocator.get_memory(e,t);return this.allocator.free(e,t),new Worker(new TextDecoder().decode(o),{type:1===Atomics.load(r,3)?"module":"classic"})},s=()=>{let e=Atomics.load(r,4),t=Atomics.load(r,5),o=this.allocator.get_memory(e,t);return this.allocator.free(e,t),JSON.parse(new TextDecoder().decode(o))};switch(Atomics.load(r,0)){case 1:{let e=o(),t=s(),i=this.assign_worker_id();console.debug("new worker "+i),this.workers[i]=e;let{promise:a,resolve:n}=Promise.withResolvers();e.onmessage=async r=>{let{msg:t,animal_id:o}=r.data;if("ready"===t&&(void 0!==o&&this.animal_workers.set(o,e),n()),"done"===t&&(this.workers[i].terminate(),this.workers[i]=void 0,void 0!==o&&this.animal_workers.delete(o),console.debug(`worker ${i} done so terminate`)),"error"===t||"exit"===t){this.workers[i]=void 0,void 0!==o&&this.animal_workers.delete(o);let e=0;for(let r of this.workers)void 0!==r&&(r.terminate(),"error"===t?console.warn("wasi throw error but child process exists, terminate "+e):console.debug("wasi exit but child process exists, terminate "+e)),e++;void 0!==this.start_worker&&(this.start_worker.terminate(),"error"===t?console.warn("wasi throw error but wasi exists, terminate wasi"):console.debug("wasi exit but wasi exists, terminate wasi")),this.workers=[void 0],this.start_worker=void 0,this.animal_workers.clear();let s=new Int32Array(this.lock,12);if("error"===t){let e=r.data.error,t=serialize(e),[o,i]=await this.allocator.async_write(new TextEncoder().encode(JSON.stringify(t)),this.lock,4);if(0!==Atomics.compareExchange(s,0,0,1)){console.error("what happened?"),this.allocator.free(o,i);return}1!==Atomics.notify(s,0)&&(console.error(e),this.allocator.free(o,i),Atomics.store(s,0,0))}else{let e=r.data.code;if(0!==Atomics.compareExchange(s,0,0,2))return void console.error("what happened?");Atomics.store(s,1,e),0===Atomics.notify(s,0)&&Atomics.store(s,0,0)}}},e.postMessage({...this.override_object,...t,worker_id:i,worker_background_ref:this.ref()}),await a,Atomics.store(r,0,i);break}case 2:{console.debug("start worker create"),this.start_worker=o();let e=s();this.start_worker.onmessage=async e=>{let{msg:r,animal_id:t}=e.data;if("ready"===r&&void 0!==t&&this.animal_workers.set(t,this.start_worker),"done"===r){let e=0;for(let r of this.workers)void 0!==r&&(r.terminate(),console.warn("wasi done but worker exists, terminate "+e)),e++;this.start_worker.terminate(),this.start_worker=void 0,void 0!==t&&this.animal_workers.delete(t);let r=new Int32Array(this.lock,12);Atomics.store(r,0,3),0===Atomics.notify(r,0)&&Atomics.store(r,0,0),console.debug("start worker done so terminate")}if("error"===r||"exit"===r){let t=0;for(let e of this.workers)void 0!==e&&(e.terminate(),"error"===r?console.warn("wasi throw error but worker exists, terminate "+t):console.debug("wasi exit but worker exists, terminate "+t)),t++;void 0!==this.start_worker&&(this.start_worker.terminate(),"error"===r?console.warn("wasi throw error but wasi exists, terminate start worker"):console.debug("wasi exit but wasi exists, terminate start worker")),this.workers=[void 0],this.start_worker=void 0,this.animal_workers.clear();let o=new Int32Array(this.lock,12);if("error"===r){let r=e.data.error,t=serialize(r),[s,i]=await this.allocator.async_write(new TextEncoder().encode(JSON.stringify(t)),this.lock,4);if(0!==Atomics.compareExchange(o,0,0,1)){console.error("what happened?"),this.allocator.free(s,i);return}if(0===Atomics.notify(o,0))throw this.allocator.free(s,i),Atomics.store(o,0,0),r}else{let r=e.data.code;if(0!==Atomics.compareExchange(o,0,0,2))return void console.error("what happened?");Atomics.store(o,1,r),0===Atomics.notify(o,0)&&Atomics.store(o,0,0)}}},this.start_worker.postMessage({...this.override_object,...e,worker_background_ref:this.ref()});break}case 3:{console.debug("terminate all workers");let e=0;for(let r of this.workers)void 0!==r&&(r.terminate(),console.debug("terminate worker "+e)),e++;void 0!==this.start_worker&&(this.start_worker.terminate(),console.debug("terminate start worker")),this.workers=[void 0],this.start_worker=void 0;break}case 4:{let e=Atomics.load(r,1);console.debug("kill animal "+e);let t=this.animal_workers.get(e);void 0!==t?(t.postMessage({msg:"kill_animal",animal_id:e}),this.animal_workers.delete(e)):console.warn(`animal ${e} not found`);break}case 5:console.debug("destroy worker background"),this.destroy()}Atomics.store(e,1,1),Atomics.store(e,2,0),0===Atomics.notify(e,2,1)&&console.warn("notify failed, waiter is late")}}};globalThis.onmessage=e=>{let{override_object:r,worker_background_ref_object:t}=e.data;worker_background=WorkerBackground.init_self(r,t),postMessage("ready")};';
 
   const blob = new Blob([code], { type: "application/javascript" });
