+++ threads/src/worker_background_worker.min.d.ts
@@ -1,0 +1,9 @@
+/**
+ * The entry point for the background worker coordination layer.
+ * 
+ * This function initializes and starts the background worker responsible
+ * for managing thread spawning and synchronization in the WASI environment.
+ */
+declare function worker_background_worker(): void;
+
+export default worker_background_worker;
